Oh well, had to go initramfs' all the way
Why? Because it's _impossible_ to uniquely identify a CD with
the root= parameter of the kernel. FFFFFFFFFFFFFFFFFF...

This was a case of "this will be simple and fast" and ....
no more ghetto shell scripting for me today.

But it works. And i think i can toss my squashfs manglers...
Maybe tinyfy bootbb a bit more. It's cute.