Apparently you need to be an expert to have the kernel boot on < 32MB RAM
Non-relocatable kernel. Startup address @16MB. Yeah that'll do it.
Now loading the kernel @ 6M, for 16M ram it'd work something like:
6 - 1 = 5MB for kernel compressed
16 - 6 = 10MB for kernel uncompressed + initramfs
... not sure if it works exactly like that, but something like that,
It has enough fluff regardless :)

The lowest RAM that worked before this was 25 MB...
Now you can start the kernel with 16MB -- but the userspace
will not really work with less than 24MB or so...
20MB gets you to the prompt eventually-ish.