Last touchies on the noninteractivity of hush problem for a bit
Just, i need to do other things ... i've got ideas tho:
do SHFLAGS=$- and add i to it if tty -s indicates so,
test for $SHFLAGS != *i* in profile.d files.
Add the gentoo ability to change terminal title with PS1.

Patch bash to look at /etc/bash/bashrc
Patch hush to have ~/.hushrc and /etc/hushrc,
and point those both and fancyps1.sh

Maybe finally, patch hush to atleast have i in $- if it's _going to_
run in interactive mode... and get rid of the $SHFLAGS hack...
or maybe go direct to that... or maybe ask busybox ML about that

Note on why testing PS1 for noninteractive is less good than $-:
It fails if PS1 is exported from an interactive shell
into the noninteractive one (tho this is not that much of
an issue for like scp/rcp since those dont have a parent
shell)