focus the boot-bb config a little bit
Add a very basic (but pretty) ls.
Because there's not _hard_ limit on the initramfs size
(well, there is, but busybox is not going to hit it, it's just
better when smaller.), i felt it was a bit elitist to expect people
to know how to use "echo *" (and/or tab completion) for ls.
I mean, the _core_ point is that the initramfs is not supposed to be
a place where you have human interaction, but ...
I've learned it'll happen regardless.

To balance it out, I also looked at the initramfs scripts vs
bootbb and dropped some unused utilities/features.